Split EventSourcedAggregateRoot from AggregateRoot.
