+++ src/Qowaiv.DomainModel/EventSourcing/IEvent.cs
@@ -1,4 +1,4 @@
-﻿namespace Qowaiv.DomainModel.Events
+﻿namespace Qowaiv.DomainModel.EventSourcing
 {
     /// <summary>The (empty) contract of an event.</summary>
     public interface IEvent { }
